Refactored user, calendar, and plants routes in backend
