+++ backend/server.js
@@ -8,7 +8,7 @@ const PORT = '8080';
 const { databaseConnection } = require('./database/mongooseConnect.js');
 const { findDocuments, findOneDocument } = require('./database/findDocuments');
 const { insertTestData } = require('./database/mockData/mockDatabase');
-const { register, login, authenticateUserRequest } = require('./auth.js');
+const { register, login, authenticateUserRequest } = require('./api/auth.js');
 const { USERS } = require('./database/models/users.js');
 
 function main() {
@@ -93,98 +93,7 @@ app.get('/mytips', (req, res) => {
   });
 });
 
-app.get('/myplants', (req, res) => {
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      findOneDocument('Users', { userId }).then((userDoc) => {
-        let myPlantsByID = [];
-        if (userDoc.myPlantsByID) {
-          myPlantsByID = userDoc.myPlantsByID;
-        }
 
-        const idProp = '_id';
-        const imageMappedPlants = [];
-        for (let i = 0; i < myPlantsByID.length; i += 1) {
-          const plant = myPlantsByID[i];
-          if (plant.image.base64) {
-            // Rewrite plant to refer to server image URL
-            delete plant.image.base64;
-            plant.image.sourceURL = `${req.protocol}://${req.get('host')}${req.originalUrl}${plant[idProp]}/image.jpg`;
-            plant.image.authenticationRequired = true;
-          }
-
-          // Rename '_id' to 'instanceID'
-          plant.instanceID = plant[idProp];
-          delete plant[idProp];
-
-          imageMappedPlants.push(plant);
-        }
-
-        res.json(imageMappedPlants);
-      });
-    } else {
-      res.status(403).json({});
-    }
-  });
-});
-
-app.get('/myplants/:plantID/image.jpg', (req, res) => {
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      const { plantID } = req.params;
-      findOneDocument('Users', { userId }).then((userDoc) => {
-        let myPlantsByID = [];
-        if (userDoc.myPlantsByID) {
-          myPlantsByID = userDoc.myPlantsByID;
-        }
-
-        const idProp = '_id';
-        const plant = myPlantsByID.find((cur) => cur[idProp].toString() === plantID.toString());
-        if (plant && plant.image.base64) {
-          res.json(plant.image.base64);
-        } else {
-          res.status(404).send();
-        }
-      });
-    } else {
-      res.status(403).json({});
-    }
-  });
-});
-
-app.put('/myplants/:plantID/image.jpg', (req, res) => {
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      const { plantID } = req.params;
-      const { base64 } = req.body.image;
-      findOneDocument('Users', { userId }).then((userDoc) => {
-        let myPlantsByID = [];
-        if (userDoc.myPlantsByID) {
-          myPlantsByID = userDoc.myPlantsByID;
-        }
-
-        const idProp = '_id';
-        const idx = myPlantsByID.findIndex((cur) => cur[idProp].toString() === plantID.toString());
-        if (idx !== -1) {
-          myPlantsByID[idx].image = {
-            base64,
-          };
-
-          USERS.updateOne({ userId }, { myPlantsByID }).then(() => {
-            res.status(200).json({});
-          }).catch((saveError) => {
-            console.error(`Failed to update a plant picture for user ID ${userId}. Reason: ${saveError}`);
-            res.status(500).json({});
-          });
-        } else {
-          res.status(404).send();
-        }
-      });
-    } else {
-      res.status(403).json({});
-    }
-  });
-});
 
 app.get('/savedplants', (req, res) => {
   authenticateUserRequest(req, res).then((userId) => {
@@ -196,189 +105,6 @@ app.get('/savedplants', (req, res) => {
         }
 
         res.json(savedPlantsByID);
-      });
-    } else {
-      res.status(403).json({});
-    }
-  });
-});
-
-app.post('/myplants', (req, res) => {
-  const idProp = '_id';
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      const query = { userId };
-      const {
-        plantID,
-        name,
-        location,
-        image,
-      } = req.body;
-      findOneDocument('Users', query).then((userDoc) => {
-        let myPlantsByID = [];
-        if (userDoc.myPlantsByID) {
-          myPlantsByID = userDoc.myPlantsByID;
-        }
-
-        // Only store known properties (don't allow client to store arbitrary data)
-        const plant = {
-          plantID,
-          name,
-          location,
-          image,
-        };
-        myPlantsByID.push(plant);
-
-        USERS.updateOne(query, { myPlantsByID }).then(() => {
-        // Refresh to find the instanceID
-        // TODO: Optimize this. We need the _id that was assigned to the added plant.
-          findOneDocument('Users', query)
-            .then((newUserDoc) => {
-              const addedPlant = newUserDoc.myPlantsByID[newUserDoc.myPlantsByID.length - 1];
-              addedPlant.instanceID = addedPlant[idProp];
-              delete addedPlant[idProp];
-              console.log(`Added plant ${plantID} to user ID ${userId}'s owned plants`);
-              res.status(201).json(addedPlant);
-            })
-            .catch((refreshError) => {
-              console.error(`Failed to add a plant to 'myplants' for user ID ${userId}. Reason: ${refreshError}`);
-              res.status(500).json({});
-            });
-        }).catch((saveError) => {
-          console.error(`Failed to add a plant to 'myplants' for user ID ${userId}. Reason: ${saveError}`);
-          res.status(500).json({});
-        });
-      });
-    } else {
-      res.status(403).json({});
-    }
-  });
-});
-
-app.post('/savedplants', (req, res) => {
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      const query = { userId };
-      const {
-        plantID,
-        plantName,
-        image,
-      } = req.body;
-      findOneDocument('Users', query).then((userDoc) => {
-        let savedPlantsByID = [];
-        if (userDoc.savedPlantsByID) {
-          savedPlantsByID = userDoc.savedPlantsByID;
-        }
-
-        // Only store known properties (don't allow client to store arbitrary data)
-        const plant = {
-          plantID,
-          plantName,
-          image,
-        };
-        savedPlantsByID.push(plant);
-
-        USERS.updateOne(query, { savedPlantsByID }).then(() => {
-          console.log(`Added plant ${plantID} to user ID ${userId}'s saved plants.`);
-          res.status(201).json({});
-        }).catch((saveError) => {
-          console.error(`Failed to add a plant to 'savedplants' for user ID ${userId}. Reason: ${saveError}`);
-          res.status(500).json({});
-        });
-      });
-    } else {
-      res.status(403).json({});
-    }
-  });
-});
-
-app.delete('/myplants/:instanceID', (req, res) => {
-  const idProp = '_id';
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      const query = { userId };
-      const { instanceID } = req.params;
-      findOneDocument('Users', query).then((userDoc) => {
-        let myPlantsByID = [];
-        if (userDoc.myPlantsByID) {
-          myPlantsByID = userDoc.myPlantsByID;
-        }
-
-        const newMyPlantsByID = myPlantsByID.filter(
-          (cur) => cur[idProp].toString() !== instanceID.toString(),
-        );
-
-        if (newMyPlantsByID.length !== myPlantsByID.length) {
-          USERS.updateOne(query, { myPlantsByID: newMyPlantsByID }).then(() => {
-            console.log(`Removed plant with instanceID=${instanceID} from user ID ${userId}'s owned plants.`);
-            res.status(201).json({});
-          }).catch((saveError) => {
-            console.error(`Failed to remove a plant from 'myplants' for user ID ${userId}. Reason: ${saveError}`);
-            res.status(500).json({});
-          });
-        } else {
-          console.log(`Failed to remove plant with instanceID=${instanceID} from user ID ${userId}'s owned plants because the instanceID was not recognized.`);
-          res.status(404).json({});
-        }
-      });
-    } else {
-      res.status(403).json({});
-    }
-  });
-});
-
-app.put('/myplants/:instanceID', (req, res) => {
-  const idProp = '_id';
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      const query = { userId };
-      const { instanceID } = req.params;
-      const newPlantProps = req.body;
-      findOneDocument('Users', query).then((userDoc) => {
-        let myPlantsByID = [];
-        if (userDoc.myPlantsByID) {
-          myPlantsByID = userDoc.myPlantsByID;
-        }
-
-        const plantIndex = myPlantsByID.findIndex((cur) => cur[idProp].toString() === instanceID);
-
-        if (plantIndex !== -1) {
-          // Copy all valid properties in request body
-          if (newPlantProps.name) {
-            myPlantsByID[plantIndex].name = newPlantProps.name;
-          }
-          if (newPlantProps.location) {
-            myPlantsByID[plantIndex].location = newPlantProps.location;
-          }
-          if (newPlantProps.image) {
-            myPlantsByID[plantIndex].image = newPlantProps.image;
-          }
-
-          USERS.updateOne(query, { myPlantsByID }).then(() => {
-            console.log(`Updated plant with instanceID=${instanceID} for user ID ${userId}.`);
-
-            const responseObj = myPlantsByID[plantIndex];
-            // Convert base64 image (if it is) to URL
-            if (responseObj.image.base64) {
-              // Rewrite plant to refer to server image URL
-              delete responseObj.image.base64;
-              responseObj.image.sourceURL = `${req.protocol}://${req.get('host')}${req.originalUrl}/image.jpg`;
-              responseObj.image.authenticationRequired = true;
-            } // TODO: This really needs refactored!
-
-            // Rename '_id' to 'instanceID' in the response
-            responseObj.instanceID = responseObj[idProp];
-            delete responseObj[idProp];
-
-            res.status(201).json(responseObj);
-          }).catch((saveError) => {
-            console.error(`Failed to update a plant in 'myplants' for user ID ${userId}. Reason: ${saveError}`);
-            res.status(500).json({});
-          });
-        } else {
-          console.error(`Failed to update plant with instanceID=${instanceID} for user ID ${userId}.`);
-          res.status(404).json({});
-        }
       });
     } else {
       res.status(403).json({});
@@ -413,114 +139,11 @@ app.delete('/savedplants', (req, res) => {
   });
 });
 
-app.get('/mycalendar/notes', (req, res) => {
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      const query = { userId };
-      findOneDocument('Users', query).then((userDoc) => {
-        const { calendarNotes = {/* Default no notes */} } = userDoc;
-        res.json(calendarNotes);
-      });
-    } else {
-      res.status(403).json({});
-    }
-  });
-});
 
-app.post('/mycalendar/notes/', (req, res) => {
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      const query = { userId };
-      const keys = Object.keys(req.body);
-      const values = Object.values(req.body);
-      const when = keys[0];
-      const note = values[0];
-      findOneDocument('Users', query).then((userDoc) => {
-        let calendarNotes = {};
-        if (userDoc.calendarNotes) {
-          calendarNotes = userDoc.calendarNotes;
-        }
-
-        if (calendarNotes[when]) {
-          // Append to the existing date's notes
-          calendarNotes[when].push(note);
-        } else {
-          // Create a new date:note pair
-          calendarNotes[when] = [note];
-        }
-        USERS.updateOne(query, { calendarNotes }).then(() => {
-          res.status(201).json({});
-        }).catch((saveError) => {
-          console.error(`Failed to post a note (${JSON.stringify(note)}) to the calendar for user ID ${userId}. Reason: ${saveError}`);
-          res.status(500).json({});
-        });
-      });
-    } else {
-      res.status(403).json({});
-    }
-  });
-});
-
-app.get('/plants/', (req, res) => {
-  const plantsQuery = {};
-  findDocuments('Plants', plantsQuery).then((docs) => {
-    res.json(docs);
-  });
-});
-
-app.get('/plants/:plantID', (req, res) => {
-  const userQuery = { plantID: req.params.plantID };
-  findDocuments('Plants', userQuery).then((docs) => {
-    if (docs.length < 1) {
-      res.status(404).json({});
-    } else {
-      res.json(docs[0]);
-    }
-  });
-});
-
-app.post('/register/', (req, res) => {
-  const {
-    username, password, firstName, lastName,
-  } = req.body;
-  register(username, password, firstName, lastName).then((status) => {
-    res.json(status);
-  }).catch((reason) => {
-    console.error(`Failed to register a user due to an error: ${reason}`);
-    res.json({});
-  });
-});
-
-app.post('/login/', (req, res) => {
-  const { username, password } = req.body;
-  login(username, password).then((status) => {
-    res.json(status);
-  }).catch((reason) => {
-    console.error(`Failed to login a user due to an error: ${reason}`);
-    res.json({});
-  });
-});
-
-app.get('/login_info', (req, res) => {
-  authenticateUserRequest(req, res).then((userId) => {
-    if (userId) {
-      findOneDocument('Users', { userId }).then((user) => {
-        res.json({
-          userId,
-          username: user.username,
-          firstName: user.firstName,
-          lastName: user.lastName,
-        });
-      });
-    } else {
-      // Not logged in, send null info
-      res.json(null);
-    }
-  }).catch((error) => {
-    console.error(`Failed to get login info due to error: ${error}`);
-    res.json(null);
-  });
-});
+require('./routes/myplants')(app);
+require('./routes/mycalendar')(app);
+require('./routes/plants')(app);
+require('./routes/users')(app);
 
 app.listen(PORT, HOST, () => {
   console.log('Listening...');
